STELLA FastWavesUV stencil
First implementation of the FastWavesUV STELLA stencil. The
implementation is still not functioning; the test crashes during the AST
analysis and the problem seems to be due to the fact that currently
assignments to local variables (e.g., `rhou' variable in stage_uv()
stencil stage) are not supported throwing a `NameError' during stage
analysis. Some other minor issues were fixed as well:

* StageBody stores the user-readable name of the stage, in order to
  print more informative messages in case of errors.
* Added support for plotting the stencil data dependencies into a file.
* Fixed the calculation of the interior domain points.