Reset internal kernel method on stencil analysis failure
If the stencil analysis fails for an unsupported Python feature in the kernel, the internal _kernel() method has already been replaced by the StencilInspector, that was able to correctly process the stencil.
This commit resets the internal kernel reference to a previously valid or default (ie raising NotImplementedError) function if the abovementioned analysis fails.