Raise exception if stencil data field depends on itself
According to gridtools' data dependency rules, we cannot write into a field that is accessed in an extent different than <0,0> (1D) in a previous stage.
Thus, a field cannot depend on itself if its access extent is different from <0,0> in all dimensions.

This commit adds a StencilScope method to check data dependencies. It should be called after Stencil.generate_code(), when all data dependencies have been resolved at stencil's scope.
For the time being, no kind of self-dependency is supported.