Added automatic stage detection from get_interior_points functions
In order to use the newly computed ghost cells when generating interior points for the Python backend, get_interior_points() functions have to know which Stage object's ghost cell to pass to the points generator.
To do this, a number of techinques to detect which stage a given call to get_interior_points() corresponds to.
* For a stage defined with a function, the name of the stage can be obtained by looking into the calling stack frame of get_interior_points
* For a stage defined through a for loop in the kernel, the position of the get_interior_points call in the kernel source code is compared to the code stored in the StencilScope and to the line numbers of Stage AST nodes.
* For a stage in a procedural stencil (which uses the static version of get_interior_points), a reference to the Stencil object is obtained by looking at the arguments of the kernel wrapper; after that, a procedure similar to the one used for for-loop-defined stages is applied.

Also added a Stencil function to return a stage given a partial name string.