fixes 394
allows the user to extract a storage.
NOTE that this follows the same pattern used when a storage is passed
from an external library, i.e. the new storage does not take ownership
on the data. If the user changes the storage (i.e. clones it back and
forth from-to the GPU) might cause strange side effects (the data field
and its snapshots being in different states).