Fixing the bug with conditional switches (#507)
Bug description: For the conditional switches with CUDA, in RELEASE mode, with GCC 5.3, the default case is selected, regardless of the conditional.

Solution: Replace the C++11 Lambdas with functors for that specific case (only modifications internal to the library, the UI remains the same). The preprocessor macro NVCC_GCC_53_BUG has been added.