Fixes copy_into_set metafunction (#484, #485)
Bug description: The meta function copy_into_set did not append to the set but only return a set of the given elements, i.e. it was appending to an empty list. See #485 for a detailed description of the side-effects of this bug.

Solution: Trivial. Append to the given set.

Fixes #485