Update performance numbers due to improvements in mergeDG branch (#480)
Description: Update the performance numbers to reflect the performance improvements introduced in the mergeDG branch.

Performance changes: Removed unnecessary __syncthreads(). The __syncthreads() is added between all stages unless the stages are inside a make_independent construct. There is also no __syncthreads() after the last stage. This mechanisms got broken in a previous merge and was restored in the merge_DG branch.
The examples affected are in particular the ones using the make_independent construct (e.g. horizontal diffusion)