Fix vertical size of temporaries (#490, GC-14)
Bug description: To determine the k size of temporaries the value of the last splitter was used (the function value_at_top() was used to retrieve this value). The offsets in the k interval where not taken into account. 

Solution: Replace the value_at_top call by a function that calculates the size of the k interval. I decided to keep the change small and use the axis to determine the size in the new function k_total_length.

Drawback: The allocated size of the temporaries might be too large if the size of the axis does not match the size of the intervals in the computation, see #491.

Additional changes:
* removed value_at_top and value_at_bottom as its intention was not really clear to me (it returned the value first and last value of the value_list).
* replaced them by k_max and k_min which calculate the first and last k level by taking into account the axis offsets
* removed code duplication between common_grids/grid_cg.hpp and structured_grids/grid.hpp, moved most of the code to the base class.
* removed the grid ctor which was taking a gridtools::array (was only used once in structured_grids) and added a delegating ctor in icgrid to be removed later.
* rename common_grids/grid_cg.hpp to grid_base.hpp following the usual pattern.

Performance changes:
* In the refactoring the partitioner was moved to the base class leading to an increase in size of the grid in icosahedral. This resulted in a performance loss of ~5 % for one of the stencils. See issue #524. We accepted the loss for now as the partitioner belongs to the base class (if it belongs to the grid at all) and the performance loss is only for one stencil.