Stabilize CPU benchmark results and update numbers (#552,GC-62)
Bug description: The CPU benchmark results fluctuate quite a lot. In order to get comparable results for CPUs the numbers have to be stable and reproducible.

Solution: Pin the threads to CPU 0..n (where n is OMP_NUM_THREADS - 1) and use exclusive mode in slurm. This might have a performance penalty but the numbers are stable.