Redesign of the accessor (#545,GC-97)
Bug Description: Together with the expandable parameters functionality, the accessor saw a redesign that was difficult to read and follow due to a complex hierarchical design and hidden features
The current design looks like

accessor : accessor_mixed< accessor_impl<...> > : accessor_base

so that a "normal" accessor became a degenerated case of an accessor. At first sight, this seemed to be required in order to equip the normal accessor with a

template<ushort_t I> int get_constexpr()

However since the accessor_mixed used to build the degenerated accessor does not contain any <Pair...>,
any use of get_constexpr() should trigger a compile time error. In practice it does not trigger any error, since the accessor_mixed builds a non empty offset_tuple_mixed, non initialized, but constexpr. So it returns always 0 to any call to get_constexpr(). This is the hidden feature that I try to make here more explicit by:

1) recovering the old design, that is more flat and easy to follow,
2) explicitly provide an API with get_constexpr in the normal accessor that returns 0