Add -march=native to build options (#582)
Feature description: Adds -march=native to the host compilation.

Performance changes: icgrid and strgrid performance numbers slightly improved by this option.

Comments: We could not reproduce the performance of the reference of the master before this commit for the copy_stencil (performance improvement of the reference with this commit) and advection_pd_bott (performance worse than previous reference). The advection_pd_bott timers are still broken in this commit and will be updated soon.