Improve error message arg not in aggregator (GC-23, #521, #421)
Bug description: Providing an error message to the user that `make_stencil` uses placeholders that are not listed in the `aggregator`. If this happened before this fixes, the error message was raised deep down the `local_domain`. Fixes #421.

Details: The static assert has been added, but it may show up after the error in the `local_domain`, which has been checked also with a static assert. The error in `local_domain` warns the user that they should be looking for other errors.