Allowing the expand_factor to be larger than the expandable parameter size (GC-122, #576, )
Feature description: The case in which a user defined an expandable parameters vector is smaller than the compile-time expand_factor value was considered an error (because there is no performance benefit in using expandable parameters in such cases). However this constraint has been relaxed, since the parameters length is a runtime parameter and may end up being less than the expand factor. In the latter case the execution happens as if the computation was replicated several times.

Detailed description:
If the parameters vector is larger than the expand_factor, then I need the m_intermediate, which contains the chunk of storages, so I instantiate it. If the size of the parameters vector is not divided exactly by the expand_factor, then I need also the m_intermediate_remainder, and I instantiate it as well (otherwise it doesn't get instantiated). Now if I allow the parameters vector to be smaller than the expand_factor, then I don't need to instantiate the m_intermediate, because I'll always use the m_intermediate_remainder. That's what these new "if" in intermediate_expand.hpp do. In the latter case the m_intermediate unique_ptr is left NULL.