Fix overload for fabs/abs and add unittest (GC-157, #591)
Bug description: The functions fabs() and abs() in gt_math did not provide the same overloads as the std library. Some overloads did not exist in device mode and a host function was taken which resulted in a warning. Fixes #415.

Solution: Add the same overload that the std library provides. Emit a compile-time error in case of long double on device. Add a unittest including checks for correct return type.

Addtional changes: Adds a cuda test wrapper which "automatically" copies the test result back to the host.