Default interval (GC-31, #466, #489)
Feature description: Implementation of an alternative API of the Do() method for the case in which the interval is always the whole k region. In this case the second argument to the Do() can be omitted, and the iteration space will automatically span the whole axis used when instantiating the grid. Fixes #466.

Detailed description:
- added a sfinae method to check whether a static member function has one or two arguments
- added an extra type wrapping the user functor, and dispatching to the correct Do method (with 1 or 2 arguments)
- added the same machinery for stencil function, for using "call" without passing a k interval.