Fix CXX11_ENABLED/CUDA8 guard in iterate_domain (GC-212, #627)
Bug description: For CUDA version < 8 the cxx03 version of the iteration_domain was used instead of cxx11.

Details: The compilation of iterate_domain_cxx11 was protected under some CUDA8 ifdefs, since it contains asserts which do not compile in cuda unless we use cuda8. But this was forcing the use of iterate_domain_cxx03 (which should be dead code) even when compiling in cxx11 mode if cuda8 is not used. Here we push the guards inside the iterate_domain_cxx11 in order to use that class even for cuda < 8.