Resolve race condition when using ij-caches (GC-197, #617)
Description: There is a race condition when using IJ cached fields in combination with multi stage stencils and offsets.

Details: The problem appears because in the last stage there is no syncthreads call but it can be that one warp is consuming values that were produced by another warp. Without synchronization it can happen that the producing warp is already in the next k-level but the old values were not consumed yet.

Performance changes: Fixing this issue introduces an additional syncthreads call after the last ESF. This of course gives a performance penalty. But due to the reason that this synchronization is simply missing we have to fix this and accept the performance loss.