Initialize advection pdbott example with different storages (GC-216, #661)
Bug description: There is a unintended test scenario in advection_pdbott_prepare_tracers. The expandable parameter (std::vector) is initialized with multiple copies of exactly the same data_store. This is clearly not intended because it should be different storages.

Performance changes: Fixing this issue increases the runtime of the advection_pdbott_prepare_tracers test by up to 20% in GPU case. The CPU tests results are much worse than before (up to 5x increase in runtime).