Enable build on compute nodes of Piz Daint (#669)
Description: Add an option to the build script to allow compilation on the compute nodes of Piz Daint.

Details: The following scripts are now run on compute nodes, i.e. scheduled in slurm
- make (1 node)
- test single node (1 node)
- test mpi (4 nodes)
- test mpi CUDA (4 nodes)
Tests have shown that compilation on the hybrid nodes is much faster compared to the multicore nodes, therefore even compilation is on hybrid nodes.