Linking issue in clonable_to_gpu (GC-102, #610, #447)
Bug description: Classes inheriting from clonable_to_gpu have different members (the CUDA version has an extra gpu ptr) and therefore different sizes when compiled with nvcc or a host compiler. However, it is possible to link these two objects together without error since the symbols are identical. This will result in weird runtime errors. Fixes #447.

Solution: The solution is to compile clonable to gpu objects in the gpu code, and declare the object as extern in the host source. Then the symbol is resolved at linking time correctly.