icgrid: Add location protection and add location type to examples (GC-114, GC-116, #624, #559)
Bug description: Currently there is no protection if using accessors on corresponding args with different location type in irregular grids. Add location_type in examples. Fixes #559.

Description: The check is done in esf.hpp so it is performed when the computation is made. The error is reported to the user. Also, a previously internal error has been translated into a user error since it checked something coming directly from the user.

Additional changes: Some examples had mistakes in the location types.