Replace operator! by global accessor (GC-67, #563, #495)
Feature description: Having storages which are global to the whole iteration space, and being able to access them from whithin the user functor with the operator(), was something achieved by prefixing an operator! to the accessor. Since there are no use cases for this when we access regular storages, and since this feature is used only for special storages which are always accessed using "global" indices, now the global_accessor is used instead of the operator().
The syntax changes because now instead of using a regular accessor we have to use a global_accessor, and there's no need anymore for the operator!, because the type of accessor is different. Fixes #495.

Detailed description:
A new type global_accessor_with_arguments is available (invisible to the user). There is an overload for operator() in the iterate_domain, so that it can be called from the Do method passing global indices like a storage (see example extended_4D for an example).

Additional changes:
Simplified the expressions metafunctions