k-caches for structured grids (GC-220, #671, #454, #525)
Feature description: Provide an implementation for k-caches. At the moment the implementation is only for the CUDA backend and structured grids. Fixes #454.

Details: the k-cache provides (if specified by the user) a local register ring buffer, private to each thread, that is used to cache values of a field accessed with vertical offsets which are reused at the execution of consecutive vertical planes. The CUDA backend implements functionality to slide the ring buffer everytime the iterators are moved in the vertical dimension, and fill & flush in order to synchronize the cache with the main memory of the gpu. 
        
Performance changes: vertical_advection_dycore is now using k-caches, which improves its performance.

API changes: 
All extents must provide correct values for the k accesses (in order for k caches to properly work). Before the k-extents where not used but could be provided. There is no check that imposes that: this will be done in another task.

Additional changes:
The check that all output parameters have a null extent is relaxed to only the IJ plane, i.e. we allow to write at non-zero offset in k. Fixes #525.