Convention for storage size and view.operator()(...) (GC-229, #688)
Feature description: operator()(0,0,0) points to first grid point in halo. The halo is kept in the storage_info type but it does not extend the storage anymore. The sizes that are passed to the storage info constructor must include the halo. 

API changes: Additionally to the dim, unaligned_dim, stride, unaligned_stride following new methods were added to the storage_info class:
1. begin: index of first non-halo data point
2. end: index of last non-halo data point
3. length: number of data points (excludes padding, halo, etc.)
4. total_begin: index of first halo point
5. total_end: index of last halo point
6. total_length: number of data+halo points (excludes padding)
7. padded_total_length: total length of the storage (halo, data, padding, initial offset)