Improve indexing for temporaries (GC-241, #700)
Feature description: In the iterate domain we are using a pointer and an offset to access a data point. The pointer used in the old indexing approach was the first accessible point of the storage (e.g., halo or data point). In order to be aligned even when using temporaries with different halo accesses in the aligned dimension the halo access information was passed as template information. This information was used to provide aligned data accesses for all temporaries but unfortunately this overhead increased compile times a lot. The new indexing is initializing the pointer to the first aligned data point (non-halo point) and provides aligned accesses for all temporaries without the need of passing plenty of type information across the gridtools internals.

Documentation: The algorithm is explained in detail in the Documentation branch (./docs/internal/new_indexing.md)

Performance changes: None