Runtime storage_info (without type information) (#699)
Feature description: Provides a run time storage_info with array of dims and strides. Fixes #697.

Details: For the Fortran interface and the interface to Serialbox it is advantageous to access information of the storage_info without type information and as arrays (such that the user does not have to write his own compile time loop over the templated functions). 

Additional changes: Provides gridtools::array to std::vector conversion.