Expression unit tests (#708)
Bug description: Expression were lacking a unit test and some expressions were broken. Due to a bug in CUDA < 8 expressions are only fully supported with CUDA8. Fixes #419

Solution:
- Fixes the following expression: -arg (negate), +arg, float-arg, float/arg and combinations of args with integral types.
- Add unit tests (some disabled for CUDA < 8).

Additional changes:
Introduces a gtest addon which allows to write CUDA_TEST to generate a test for host and a test for device. The code in the test body is executed in a kernel. Values which are inside a ASSERT_EQ and ASSERT_FLOAT_EQ are copied back to host and compared there. All other ASSERTS and EXPECTS are not supported. Only bool, int and float is supported as values to be compared.