Update communication module (GC-214, #630)
Description: Reworked some communication code that broke when testing COSMO-like decomposition with one-sided communication. The problem was in a implicit reshaping performed by the processor grid that lead to wrong results. The solution has been to split the constructors that take extra dimensions and the one accepting just a communicator, which is assumed to be then a cartesian one. The dimensions are treated as in MPI_Dims_create, so that the values different than 0 are modified, while the others remains untouched. Also the All-to-all pattern has been fixed to be decreasing stride oriented.

Additional changes: CUDA_VERSION has been changed into GT_CUDA_VERSION since cuda.h on kesch defined its own value that made the code to wrongly pick CUDA8.