Abstract reassign domain in aggregator (GC-225, #725)
Feature description: Enables possibility to use of reassign domain in cases were type erasure is needed.

Details: With the previous implementation of the reassign domain the full type of the computation has to be retained. This is not practical for large codes as compile time would explode (the type would need to be in a header and would be compiled for every translation unit).
The solution in this PR is to partly type-erase such that the placeholders are kept and the reassign works on this partly type-erased. The fully type-erased structs are called stencil and reduction (former computation) and the new partly type-erased struct is computation<Aggregator, ReductionType>.

API changes: Additional overload of run() which accepts parameters with the same interface as reassign (as a short-cut for reassign() + run()), i.e. either just a complete ordered list of storages or arg-storage pairs (only storages which should be reassigned).