Fix for distributed examples (GC-15, #713)
Description: Fixed `copy_stencil_parallel` and `shallow_water_enhanced`.

Details:
The examples are implemented without the partitioner. A new API for accessing view information is added which uses friend functions to access private information of the view in a somewhat cumbersome way. These are put in a a`gridtools::advanced` namespace.
Shallow-water-enhanced is only compiling since reference and gridtools implementation do not seem to be compatible. This should pick it up in the future. We decided to incorporate this PR anyway to not delay further.