Bypass extent analysis (GC-245, #727)
Feature description: This update introduces the ability to skip the extent analysis for stencils for which the user had done the extent computation already. The stages are then associated with an extent by calling `make_stage_with_extents<functor, extent>` so that the map between stages and extents to compute temporary sizes is done directly with that information.

Additional information: This is a change the is possible since now all temporaries share the same sizes to reduce register pressure. The map between placeholders and extents are not needed anymore in the backend, otherwise this interface would not work. A future architecture that requires the explicit map for performance, is going to deprecate this approach.

API change: adding `make_stage_with_extents`