Distributed boundaries (GC-270, #812)
Feature description: This is a new library in the GridTools. The description can be found in `docs/manuals/user_manual/comm_and_bc.hrst`. The main idea is to combine boundary conditions and communication in a single interface. The API will instantiate the `halo_exchange` pattern and the use is to indicate a list of "jobs", each consisting either of data stores to exchange of boundary condition specifications from which the data stores for communication will be extracted. The interface allows to leave certain data stores out from the communication if they are needed only in boundary conditions or if they are read-only.