Doxygen (GC-321, #839)
Description: Trying to reviving doxygen.

Details:
Distributed boundaries and boudary conditions are examples of how in-code documentation should be implemented. Other code has been cleaned up to make the resulting documentation more usable.

The main idea is that each folder in `include` is a "Module", that is, it is collected in a doxygen group. The main include file in the folder contains the definition of the group, and each file contains an `ingroup` directive with `@{` at the beginning and a `@}` at the end. Sub folders should be put as Sections, but I haven't done this yet. The resulting documentation at this point, when tar-gz the folder is more than 1G bytes.