Bugfix: Force inlining in GT_FUNCTION_WARNING (GC-351, #864)
Description: Forces inlining of functions that use the GT_FUNCTION_WARNING macro and added Doxygen documentation to explain the meaning of the different macros. The macro is used in places where a function is meant to be run on host only but needs the `__device__` qualifier to prevent the compiler from issuing a warning.