Macro to generate a repository of data_stores (GC-180, #717)
Feature description: Adds a macro to generate a repository of data_stores.

Details:
The macro generates a class containing
- data_stores of specified types and names with getters.
- a map<string,variant<data_store_types...>> to access these data_stores by a string
Adds a macro to generate a special variant (inheriting from boost::variant) with conversion operators to all contained types which allows automatic conversion. This variant is used in the map above.