API change: fix get_interval in k-axis API (GC-338, #854)
Bug description: get_interval<X> and get_interval<X+1> is overlapping.

Solution:
get_interval<X> will return an interval interval<level<X,+1>,level<X+1,-1>>

Details:
Previously the first splitter was placed after the first k-level and the last splitter after the last level which is asymmetric and confusing:
```x|xxxxxxx|```
Now the first splitter is to the left of the first splitter:
```|xxxxxxxx|```
In this example with 8 k-levels the elements of the `value_list` are `{0,8}` while previously it was `{0,7}`. To achieve that the 0 is not moved the `level<0,1>` is associated with `value_list[0]` while previously `level<0,-1>` was associated with `value_list[0]`.

API changes: The previously described re-definition of the `value_list` elements to the `<X,+1>` level is a tiny but very severe change: all users which are still using the old way of defining the k-axis (i.e. not using the get_interval and filling the value_list manually), will have to carefully fix their intervals. Users of the new k-axis API (or the default interval) are not affected by this change.
Therefore, it is recommended to update your code to the k-axis API before using this fix. Because of the bug described above you need to use a work-around: instead of `get_interval<X>` use `get_interval<X>::modify<1,0>` for all `X>1` (not for `X==0`).

Performance: No change.