Bug: nested call with shifted accessor index (GC-416, #882)
Bug description: When getting the accessor_return_type from the the CallerAggregator we wrongly passed the accessor of the callee, instead of the caller. The bug was undetected because the value was correct, only the type was wrong, therefore it worked in many cases.

Details: Fixed the bug and added a separate (compile-time) test that ensures that types are correctly passed between stages. Added a missing specialization for the remapper in case of a wrapped reference (when a local variable is passed).