Layout transformations (GC-318, #885)
Description: Introduces a function to apply layout transformations, taking src and dst pointers, shapes and strides. Executes the transformation on host/device by automatically detecting the ptr type.

Details: As the function is intended to be used in a generic way, e.g. from Fortran, it acts on raw ptr with dimensions and strides passed explicitly, not on data_stores.
- adds is_gpu_ptr to common
- adds join_array() to concatenate to arrays

Additional changes:
- default_layout_map<1...4> replaced by generic one for any number of dimensions