Alternative CUDA boundary stencil (GC-417, #886)
Description: This PR provides a different implementation of the boundary application for GPUs. Instead of running (potentially) 26 kernels it runs only one. The benefit for real applications needs be evaluated. This implementation offer a more flexible base to work with to improve performance of boundary condition applications. We keep the old implementation for now, as it is not yet clear if the more complicated implementation will pay off in performance.