Fix alignment and storage_info API cleanup (GC-151, #837)
Description: Alignment before this PR was implemented in a non general way. It assumed the memory allocation provided a prescribed alignment (for cuda the value is 256 bytes). This limits the amount of alignment that can be guaranteed. 

Now alignment is specified as before, by indicating that the aligned address (the beginning of the inner region indicated by `halo`) will be multiple of `Align * sizeof(storage_value_type)`. Why alignment is generally specified in bytes and not in `sizeof(value_type)`, changing this would imply that `storage_info` would need to know about the value type of the storage in which it will be used, and this is not our current design.

Additional changes: This PR also cleans up the interface to `storage_info` since some member functions are not needed, and some computations (`m_initial_offset`) is not needed anymore in the `storage_info`.

API changes:
- `storage_info::get_initial_offset()` , `storage_info::begin()` `storage_info::end()`, `storage_info::total_begin()` `storage_info::total_end()` are removed [initial offset would be always `0`, `begin()` and `end()` would not give any indication of the iteration space, and `total_begin()` and `total_end()` are never used to iterate over data, and would iterate over padding and everything else - it was not clear from the api if initial_offset was included, too]
- `storage_info::length`, `storage_info::total_length` and `storage_info::padded_total_length` return the number of elements in the inner region, the inner_region plus halo, and the one including padding, respectively
- `storage_info::length<Dim>`, `storage_info::total_length<Dim>` and `storage_info::padded_length<Dim>` return the number of elements in the specified dimension index `Dim` for inner, outer and padded regions, respectively
- `storage_info::dim`, `storage_info::unaligned_dim`, `storage_info::unaligned_strides` have been removed, since they were ambiguous, redundant and, for the case of unaligned_strides, not clearly defined
- `cuda_storage`, `host_storage`, and `mic_storage` do not have constructors to initialize with a value anymore. There is only one constructor that takes a lambda. `data_store` can still be constructed with a value
- `cuda_storage`, `host_storage`, and `mic_storage` main constructor requires the size, the offset to align (typically the first element of the inner region), and the alignment value, in terms of number of elements of the storage
- I introduced other functions to get storage_info classes with a custom alignment, since I wanted to try them out

Performance changes: No performance changes, because our benchmarks are not sensitive to the changes in this PR.