Fix timers in expandable parameters (#906)
Description: In expandable parameters the time measurement was accumulating the timers of chunk and remainder separately. This hides performance problems in between launches of chunks. Currently the master contains such a problem which will be fixed in the next commit.

Performance changes: AdvectionPDBott performance is much worse, because we have a bug in the implementation of expandable parameters, but the measurement is now done properly. Will be fixed in the next commit.