Bugfix: return type in call (GC-454, #908)
Description: This is an addition to the fix in #882 where the `accessor_return_type` was fixed for `call_proc`. This fix revealed a similar issue in the `call`.
Our (arbitrary) convention was, that the return type of a `call` is the type of the first `inout_accessor` of the caller. However the implementation was wrong and this behavior cannot easily be fixed. Therefore this PR introduces a new convention (which is still arbitrary) to use the type of the first argument passed to the function.

API changes: Return type of `call` is the type of the first argument in the call. Additionally, the return type can be specified by `call:return_type<MyReturnType>::with(...)`.