Fix include paths (GC-357, #934)
Description: This PR fixes the include path in `CMakeLists.txt`, which before this PR included `include/gridtools`. 

Details:
This PR then fixes the `examples` and `unit_tests` to consistently use `#include <gridtools/...>`, but then it fixes also the conventions to include files from within the library files, which is to use relative paths (this convention helps in maintaining in a consistent state the structure of the overall software - an aggressive modularization may be specifies in slightly different ways, like include with absolute paths for inter-module includes).