Use find_package for all machines (#954)
Dsscription:
  Use find_package for all machines
Details:
find_package(MPI) must (and is able to) detect the compiler wrapper; this is way I see no reason to work around this (and it uncommon to do that). This should resolve the problems that we need to set manual include paths using $ENV{MPICH_DIR} which were due to the fact that we did not use the wrapper properly: The files were compiled with g++ instead of using the wrapper.

@ChristopherBignamini We will need to do the same thing for the dycore.

Note: Currently I am not aiming for a completely proper solution regarding cleaning up the build scripts / CMakeLists as this would be a separate task, but this should be a step into that direction.