Bugfix for contains_no_IJ_cache_t (#998)
Description: We compared against an exact type `boost::integral_constant<bool, true>` instead of its contained value. This caused that we wrongly assume that all stencils don't contain IJ caches which implies that we don't do a necessary synchronization. This was fixed already in #993, but the reference was not updated (i.e. performance is worse, but results are fixed; it is a race-condition which we did not catch with our tests). Here we just fix the computation of the contains_no_IJ_cache_t.

Performance changes:
Reference reverted (degradation) to before #988, because the change in reference was due to the incorrect cache behavior introduced in #980, which was again fixed in #993.