Communication: Fix leak in MPI communicators (#1033)
Description: There was a leak problem with MPI communicators. The Halo update pattern did not clean up at destruction, and this was because we were copying communicators that were created somewhere else. Since MPI implementations has a limited number of communicators that they can handle, poor implementations of user code could exhaust them, or destroy communicators before their last use. For this reason, we needed to deprecate the constructors of the halo-update pattern that build communicators, and do a better job at freeing communicators.

API change: This has the implication that the communicator given to the distributed boundaries has to be a Cartesian communicator.