CMake refactoring: modularization (GC-526, #1032)
Description: This PR is the first step in refactoring CMake files to allow for exporting of targets and custom installations. Fixes #1034, fixes #961.

Details: The first change is to use CUDA as a language, so that the same commands can be used for C++ targets and CUDA targets. This feature is still quite new in the CMake world and required quite some smithing, but should improve with new versions on CMake.
The second change was the ability to install only certain components, that is copy only the sources and libraries of, and run only the tests for, the components that the user decide to get. This does not incorporate a full featured export, but it's a first step.

Additional changes: I introduced here the test manifest, a file in the build folder, like the install-manifest, that lists all the tests executed when making GridTools. This file can be used to compare branches, to see if they run the same tests, which turned useful to find several bugs in my branch, and in master, too.