API change: Remove enumtype namespace from intent and execution (#1143)
Description: `intent` is now a scoped enum. Execution policies are moved to `execute` namespace, policies are types in this namespace. Vertical block size (for parallel CUDA execution) is moved to `execute::parallel_block`. Fixes #520.

Details: Note that the enumtype namespace still exists for the `location_type` which is not part of this PR.

API changes:
- `enumtype::in` and `enumtype::inout` -> `intent::in`, `intent::inout`
- `execute<enumtype::forward>` etc. -> `execute::forward`