CUDA: Fix unrolling in hypercube_iterator (#1196)
Details: nvcc was not able to unroll the loop in operator++ of the grid_iterator which resulted in use of stack frame with quite some performance penalty in the layout transformation. Note that the layout transformation should be implemented with compile-time-known number of dimensions which doesn't need the hypercube_iterator there.