Strong type for axis arguments (#1257)
Description: Replace the last two integer template parameters of axis<> by types encapsulating the integer value. The reason is that the currently second parameter should be considered internal API and dropped as soon as possible. The third parameter has to be kept for longer. To avoid weird behavior of user-code when upgrading, the only sane way to interchange the 2 parameters is by wrapping them in types to produce a compile-time error.

API change: Parameters of axis are encapsulated in types, order of these parameters is reversed.