fix the bug with delegating optional concept functions (#1531)
`delegate` helper was redesigned. Now it implements only those concept functions that are defined in the source SID. In the previous design all concept functions were defined (including those that are fall back to default for the source SID). As a result it was a lot of hard to debug compilation errors.