Introduce dump backend  (#1546)
Stencil `dump` backend is added. It doesn't perform calculations. Instead it outputs `JSON` representation of the spec (the type that is passed to the backend entry point). Unlike other backends `dump` is stateful -- it contains the reference to the `std::ostream` that is served as an output sink. The usage is like that:
 ```
  run(fe_spec, dump{std::cout}, ...);
```
The context of this PR is the following: It is a step to extract the cosmo dycore model from the gt2 based source and to pass it into gt4py in the future. The main showstopper on this way is the fact that the elementary stencils are represented in the generated `json` as `C++` demangled type names. The next step could be to parse the sources with the python based tool and to extract the AST's from the stencil bodies.