fix performance (#1611)
With C++17, `as_const` was added into the standard and if `as_const` is not called qualified, we end up with `std::as_const` (which is also a better match) in most cases because there is some `std` inside one of the template arguments.

By qualifying `as_const`, performance is fixed.

For GTBench, performance is identical with `-std=c++14` and `-std=c++17` now.